requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.4
Flask-Migrate==4.0.4
psycopg2-binary==2.9.9